refactor: UserActivityDto로 이름 변경
