+++ bookduck/src/main/java/com/mmc/bookduck/domain/badge/dto/common/UserActivityDto.java
@@ -1,6 +1,6 @@
-package com.mmc.bookduck.domain.badge.entity;
+package com.mmc.bookduck.domain.badge.dto.common;
 
-public record UserActivity(
+public record UserActivityDto(
         long readCount,
         long archiveCount,
         long oneLineCount,
