+++ bookduck/src/main/java/com/mmc/bookduck/domain/badge/service/BadgeUnlockService.java
@@ -4,16 +4,14 @@ import com.mmc.bookduck.domain.alarm.service.AlarmByTypeService;
 import com.mmc.bookduck.domain.archive.repository.ExcerptRepository;
 import com.mmc.bookduck.domain.archive.repository.ReviewRepository;
 import com.mmc.bookduck.domain.badge.entity.Badge;
-import com.mmc.bookduck.domain.badge.entity.UserActivity;
+import com.mmc.bookduck.domain.badge.dto.common.UserActivityDto;
 import com.mmc.bookduck.domain.badge.entity.UserBadge;
 import com.mmc.bookduck.domain.badge.repository.UserBadgeRepository;
 import com.mmc.bookduck.domain.book.entity.ReadStatus;
-import com.mmc.bookduck.domain.book.entity.UserBook;
 import com.mmc.bookduck.domain.book.repository.UserBookRepository;
 import com.mmc.bookduck.domain.oneline.repository.OneLineRepository;
 import com.mmc.bookduck.domain.user.entity.User;
 import com.mmc.bookduck.domain.user.repository.UserGrowthRepository;
-import com.mmc.bookduck.domain.user.service.UserGrowthService;
 import com.mmc.bookduck.global.exception.CustomException;
 import com.mmc.bookduck.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
@@ -36,7 +34,7 @@ public class BadgeUnlockService {
     private final AlarmByTypeService alarmByTypeService;
 
     // 유저 활동 가져오기
-    public UserActivity getUserActivity(User user) {
+    public UserActivityDto getUserActivity(User user) {
         // 각 활동 데이터 조회
         long readCount = userBookRepository.countByUserAndReadStatus(user, ReadStatus.FINISHED);
         long archiveCount = reviewRepository.countByUser(user) + excerptRepository.countByUser(user);
@@ -44,13 +42,13 @@ public class BadgeUnlockService {
         long level = userGrowthRepository.findByUser(user).orElseThrow(()-> new CustomException(ErrorCode.USERGROWTH_NOT_FOUND)).getLevel();
 
         // 결과 반환
-        return new UserActivity(readCount, archiveCount, oneLineCount, level);
+        return new UserActivityDto(readCount, archiveCount, oneLineCount, level);
     }
 
     // 뱃지 획득 트리거
     public void checkAndUnlockBadges(User user) {
         // 사용자 활동 데이터 가져오기
-        UserActivity activity = getUserActivity(user);
+        UserActivityDto userActivityDto = getUserActivity(user);
 
         // 현재 사용자가 이미 보유한 뱃지
         List<Long> ownedBadgeIds = userBadgeRepository.findAllByUser(user).stream()
@@ -66,7 +64,7 @@ public class BadgeUnlockService {
             }
 
             // 뱃지 조건 확인
-            boolean isConditionMet = isBadgeConditionMet(badge, activity);
+            boolean isConditionMet = isBadgeConditionMet(badge, userActivityDto);
 
             if (isConditionMet) {
                 // UserBadge 생성 및 저장
@@ -83,7 +81,7 @@ public class BadgeUnlockService {
     }
 
     // 뱃지 획득조건 충족여부 확인
-    private boolean isBadgeConditionMet(Badge badge, UserActivity activity) {
+    private boolean isBadgeConditionMet(Badge badge, UserActivityDto activity) {
         int unlockValue = badgeService.getBadgeUnlockValue(badge);
 
         return switch (badge.getBadgeType()) {
